[BusAux]: change engine speed from integer to double

+++ VECTOAux/VectoAuxiliaries/Pneumatics/ICompressorMap.vb
@@ -26,7 +26,7 @@ Namespace Pneumatics
         ''' <param name="rpm">compressor rotation speed</param>
         ''' <returns></returns>
         ''' <remarks>Single</remarks>
-        Function GetFlowRate(ByVal rpm As Integer) As Single
+		Function GetFlowRate(ByVal rpm As Double) As Single
 
         ''' <summary>
         ''' Returns mechanical power at rpm when compressor is on
@@ -34,7 +34,7 @@ Namespace Pneumatics
         ''' <param name="rpm">compressor rotation speed</param>
         ''' <returns></returns>
         ''' <remarks>Single</remarks>
-        Function GetPowerCompressorOn(ByVal rpm As Integer) As Single
+		Function GetPowerCompressorOn(ByVal rpm As Double) As Single
 
         ''' <summary>
         ''' Returns mechanical power at rpm when compressor is off
@@ -42,7 +42,7 @@ Namespace Pneumatics
         ''' <param name="rpm">compressor rotation speed</param>
         ''' <returns></returns>
         ''' <remarks>Single</remarks>
-        Function GetPowerCompressorOff(ByVal rpm As Integer) As Single
+		Function GetPowerCompressorOff(ByVal rpm As Double) As Single
 
         'Returns Average Power Demand Per Compressor Unit FlowRate
         Function GetAveragePowerDemandPerCompressorUnitFlowRate() As Single
